ACPI: Add functions to read *SDT and tables from memory
What is in FLASH (top 128K) and what kernels show you is not
always the same. Here is an example, the first coming from ebda, the
second from linux files in /sys:

rminnich@xcpu:~/gopath/src/github.com/u-root/u-root/cmds/exp/acpicat$ sudo ./acpicat  -s ebda | ../acpigrep/acpigrep -d . >/dev/null
2020/05/11 21:57:51 Keeping RSDT@0x0 64 1 0x5b "INTEL " "440BX   " 0x06040000 0x50544c20 0x00000000
2020/05/11 21:57:51 Keeping FACP@0x0 116 1 0xf7 "INTEL " "440BX   " 0x06040000 0x204c5450 0x000f4240
2020/05/11 21:57:51 Keeping DSDT@0x0 138866 1 0xaa "PTLTD " "Custom  " 0x06040000 0x5446534d 0x03000001
2020/05/11 21:57:51 Keeping BOOT@0x0 40 1 0xa5 "PTLTD " "$SBFTBL$" 0x06040000 0x50544c20 0x00000001
2020/05/11 21:57:51 Keeping APIC@0x0 1858 1 0x8f "PTLTD " "\t APIC  " 0x06040000 0x50544c20 0x00000000
2020/05/11 21:57:51 Keeping MCFG@0x0 60 1 0x6e "PTLTD " "$PCITBL$" 0x06040000 0x50544c20 0x00000001
2020/05/11 21:57:51 Keeping SRAT@0x0 2256 2 0xa8 "VMWARE" "MEMPLUG " 0x06040000 0x20574d56 0x00000001
2020/05/11 21:57:51 Keeping HPET@0x0 56 1 0xaa "VMWARE" "VMW HPET" 0x06040000 0x20574d56 0x00000001
2020/05/11 21:57:51 Keeping WAET@0x0 40 1 0x62 "VMWARE" "VMW WAET" 0x06040000 0x20574d56 0x00000001
rminnich@xcpu:~/gopath/src/github.com/u-root/u-root/cmds/exp/acpicat$ sudo ./acpicat  | ../acpigrep/acpigrep -d . >/dev/null
2020/05/11 21:57:56 Keeping APIC@0x0 1858 1 0x8f "PTLTD " "\t APIC  " 0x06040000 0x50544c20 0x00000000
2020/05/11 21:57:56 Keeping BOOT@0x0 40 1 0xa5 "PTLTD " "$SBFTBL$" 0x06040000 0x50544c20 0x00000001
2020/05/11 21:57:56 Keeping DSDT@0x0 138866 1 0xaa "PTLTD " "Custom  " 0x06040000 0x5446534d 0x03000001
2020/05/11 21:57:56 Keeping FACP@0x0 244 4 0x51 "INTEL " "440BX   " 0x06040000 0x204c5450 0x000f4240
2020/05/11 21:57:56 Keeping FACS@0x0 64 0 0x00 "\x00\x00\x00\x00\x00\x00" "\x00\x00\x00\x00\x00\x00\x00\x00" 0x00000000 0x00000000 0x00000000
2020/05/11 21:57:56 Keeping HPET@0x0 56 1 0xaa "VMWARE" "VMW HPET" 0x06040000 0x20574d56 0x00000001
2020/05/11 21:57:56 Keeping MCFG@0x0 60 1 0x6e "PTLTD " "$PCITBL$" 0x06040000 0x50544c20 0x00000001
2020/05/11 21:57:56 Keeping SRAT@0x0 2256 2 0xa8 "VMWARE" "MEMPLUG " 0x06040000 0x20574d56 0x00000001
2020/05/11 21:57:56 Keeping WAET@0x0 40 1 0x62 "VMWARE" "VMW WAET" 0x06040000 0x20574d56 0x00000001

The FACS is generated, somehow, but looks like garbage.
The standard method, via /sys, does not include the RSDT, and the ebda method does.

Note that the address is not available in the output stream (so you see @0x0) so acpigrep
shows 0.

Also, the use of uint64 for physical addresses was an early mistake,
cleaned up here. int64 is 63 bits. By the time we
need more than 63 bits for a physical address,
we will be using 128-bit machines, so uint64 won't help much anyway;
and int64 is compatible with Pread where uint64 is not.

+++ pkg/acpi/rsdp.go
@@ -18,7 +18,7 @@ var (
 // RSDP is the v2 version of the ACPI RSDP struct.
 type RSDP struct {
 	// base is the base address of the RSDP struct in physical memory.
-	base uint64
+	base int64
 
 	data [headerLength]byte
 }
@@ -64,7 +64,7 @@ func (r *RSDP) OEMID() string {
 }
 
 // RSDPAddr returns the physical base address of the RSDP.
-func (r *RSDP) RSDPAddr() uint64 {
+func (r *RSDP) RSDPAddr() int64 {
 	return r.base
 }
 
@@ -72,15 +72,15 @@ func (r *RSDP) RSDPAddr() uint64 {
 //
 // It will preferentially return the XSDT, but if that is
 // 0 it will return the RSDT address.
-func (r *RSDP) SDTAddr() uint64 {
+func (r *RSDP) SDTAddr() int64 {
 	b := uint64(binary.LittleEndian.Uint32(r.data[16:20]))
 	if b != 0 {
-		return b
+		return int64(b)
 	}
-	return binary.LittleEndian.Uint64(r.data[24:32])
+	return int64(binary.LittleEndian.Uint64(r.data[24:32]))
 }
 
-func readRSDP(base uint64) (*RSDP, error) {
+func readRSDP(base int64) (*RSDP, error) {
 	r := &RSDP{}
 	r.base = base
